+++ tests/ids-treemap/ids-treemap.spec.ts
@@ -1,11 +1,11 @@
 import AxeBuilder from '@axe-core/playwright';
-import { expect } from '@playwright/test';
-import { test } from '../base-fixture';
+import { Locator } from '@playwright/test';
+import { test, expect } from '../base-fixture';
 import IdsContainer from '../../src/components/ids-container/ids-container';
 import IdsTreeMap from '../../src/components/ids-treemap/ids-treemap';
 
 test.describe('IdsTreemap tests', () => {
-  let treemap: any;
+  let treemap: Locator;
   const url = '/ids-treemap/example.html';
 
   test.beforeEach(async ({ page }) => {
@@ -64,16 +64,31 @@ test.describe('IdsTreemap tests', () => {
       });
     });
 
+<<<<<<< HEAD
     test.skip('can resize the width when the viewport changes', async ({ page }) => {
       await page.setViewportSize({ width: 589, height: 9999 });
       let treemapWidth = await page.evaluate(`document.querySelector("ids-treemap").width`);
       let containerWidth = await page.evaluate(`document.querySelector("ids-treemap").container.offsetWidth`);
       expect(treemapWidth).toEqual(containerWidth);
+=======
+    test('can resize the width when the viewport changes', async ({ page }) => {
+      // removes the padding of the ids-container and return the tree map width
+      await page.evaluate(() => { document.querySelector<IdsContainer>('ids-container')!.padding = '0'; });
+      let currWidth = await treemap.evaluate((element: IdsTreeMap) => element.width);
+      expect(currWidth).toBeInAllowedBounds(page.viewportSize()!.width, 5);
+>>>>>>> 1ef865a9bc08957bd351d3e3e620d81511343143
 
-      await page.setViewportSize({ width: 989, height: 9999 });
-      treemapWidth = await page.evaluate(`document.querySelector("ids-treemap").width`);
-      containerWidth = await page.evaluate(`document.querySelector("ids-treemap").container.offsetWidth`);
-      expect(treemapWidth).toEqual(containerWidth);
+      let expWidth = 589;
+      await page.setViewportSize({ width: expWidth, height: 9999 });
+      await page.waitForFunction((width) => (window.innerWidth === width), expWidth);
+      currWidth = await treemap.evaluate((element: IdsTreeMap) => element.width);
+      expect(currWidth).toBeInAllowedBounds(expWidth, 5);
+
+      expWidth = 989;
+      await page.setViewportSize({ width: expWidth, height: 9999 });
+      await page.waitForFunction((width) => (window.innerWidth === width), expWidth);
+      currWidth = await treemap.evaluate((element: IdsTreeMap) => element.width);
+      expect(currWidth).toBeInAllowedBounds(expWidth, 5);
     });
 
     test('can render via document.createElement (append early)', async ({ page }) => {
@@ -135,13 +150,11 @@ test.describe('IdsTreemap tests', () => {
       expect(height).toBe(300);
     });
 
-    test('can set the treemap width', async ({ page }) => {
-      const tmContainer = await page.locator('ids-container').first();
-      // eslint-disable-next-line @typescript-eslint/no-unused-expressions
-      const tmWidth = await tmContainer.evaluate((treemapEl: IdsTreeMap) => { treemapEl.offsetWidth as number; });
-      // eslint-disable-next-line @typescript-eslint/no-unused-expressions
-      const width = await treemap.evaluate((treemapEl: IdsTreeMap) => { treemapEl.offsetWidth as number; });
-      expect(tmWidth).toBe(width);
+    test('can set the treemap width', async () => {
+      expect(await treemap.evaluate((treemapEl: IdsTreeMap) => {
+        treemapEl.width = 300;
+        return treemapEl.width;
+      })).toEqual(300);
     });
   });
 });